Refine the orientation
Since the dials-refined orientation seems to yield model shoeboxes that are too far away
from ground truth it seems like postrefinement is needed.  Here we will refine the
rotx and roty angles based on nanoBragg image simulation.  Rather than do the extra
work to get analytical derivatives, just calculate a finite diff derivative for the
ROI, and assume it is constant over the rotational range (up ot 0.08 degrees rotation).
This is not necessarily a good assumption.  Still need to see if this can work.