pixel_refinement updates for tz5101 production run.
Change the get_items function so that the parallel job is restartable, so
the second time around, work parcels (keys) are distributed differently to
the MPI ranks. Module random.shuffle is used for this.

The drop_convergence_test_max_drop_eps is lowered to 1.e-5.  It is believed
that this value was used for the production run instead of 5.e-5 so as get
many more iterations toward convergence.

It is believed that despite the listed workaround, we are still not using
OpenMP for the individual calls to nanoBragg_add_spots().  This needs to be
checked again.  It could be that we are drastically underutilizing cori.

LBFGS fit is put inside try/except, so that RuntimeError for one rank does not
kill the entire parallel job.

For each LBFGS iteration print verbose lineout including the xyz rotational
perturbations in units of 0.01 degree.