Suppress the add_noise function call.
This is the last commit on this branch to support the ADSE13-196 report, version 1.
The add_noise() function cannot be made portable between C++ and CUDA with the current design as it relies on Mersenne twister random numbers.  CUDA does not preserved the loop-through-x-loop-through-y pixel evaluation done by C++.  add_noise() is also not critical for future science aims (it is important for image simulation, not for inverse physical modeling), so dispense with it here.