Update the compare.py test with PHIL command line args.
A new set of reference images was calculated using CPU-only methods by the summit_make_reference.sh LSF script.
The updated compare.py was then used to assure that the GPU results reported in ADSE13-196 were compatible
with the CPU results.  This assumed that add_noise() is not used, and allows for 100 pixels to be +/- 1 unit off
due to rounding differences in the GPU vs. CPU add_background.

The new reference images are in ls49_big_data/adse13_196/report1.  The command line for image comparisons is
given in adse13_196/README.md