Test the new CUDA kernel optimized for masks.
The Exascale API can switch between the old kernel and the new one, which assumes that a mask is provided to screen out 95 to 99% of the pixels having no Bragg signal.

Also, back out previous attempt to make GPU tests portable.  It is not currently possible because the spectrum simulation takes data from the NeXus master (big file).